moving salome utilities
