--- kratos_salome_plugin/salome_dependent/README.md
@@ -1,1 +1,0 @@
-This directory contains files that depend on Salome. Hence they can only be imported if running inside Salome.
