+++ tests/test_geometries_io.py
@@ -26,7 +26,7 @@ from kratos_salome_plugin.utilities import IsExecutedInSalome
 from testing_utilities import SalomeTestCaseWithBox, CheckIfKratosAvailable
 
 if IsExecutedInSalome():
-    from kratos_salome_plugin.salome_dependent import salome_utilities
+    from kratos_salome_plugin import salome_utilities
 
 # Kratos imports
 kratos_available = CheckIfKratosAvailable()
