+++ tests/testing_utilities.py
@@ -21,7 +21,7 @@ if utils.IsExecutedInSalome():
     # Check https://docs.salome-platform.org/latest/tui/KERNEL/kernel_salome.html for how to handle study
     # imports that have dependenices on salome, hence can only be imported if executed in salome
     import salome
-    import kratos_salome_plugin.salome_dependent.salome_utilities as salome_utils
+    import kratos_salome_plugin.salome_utilities as salome_utils
 
     # initialize salome, should be done only once
     salome.salome_init()
